Revert commit of using float for Z Rotation
Reverts commit e4466ceb16895dfd3dddc6a764f79148ba05f188
Decided decimal precision is still not required on Z Rotation as the difference is imperceptible.

+++ Database/sql/createTables.sql
@@ -1,7 +1,7 @@
 --Create new tables to be imported into
 CREATE TABLE SeventySix_FormID(entityFormID TEXT, displayName TEXT, editorID TEXT, signature TEXT);
-CREATE TABLE SeventySix_Interior(referenceFormID TEXT, cellFormID TEXT, x INTEGER, y INTEGER, z INTEGER, locationFormID TEXT, lockLevel TEXT, primitiveShape TEXT, boundX INTEGER, boundY INTEGER, boundZ INTEGER, rotZ REAL, spawnClass TEXT);
+CREATE TABLE SeventySix_Interior(referenceFormID TEXT, cellFormID TEXT, x INTEGER, y INTEGER, z INTEGER, locationFormID TEXT, lockLevel TEXT, primitiveShape TEXT, boundX INTEGER, boundY INTEGER, boundZ INTEGER, rotZ INTEGER, spawnClass TEXT);
 CREATE TABLE SeventySix_NPCSpawn(npc TEXT, class TEXT, locationFormID TEXT, chance REAL);
 CREATE TABLE SeventySix_Quantified_Scrap(component TEXT, componentQuantity TEXT, junkFormID TEXT);
-CREATE TABLE SeventySix_Worldspace(referenceFormID TEXT, x INTEGER, y INTEGER, z INTEGER, locationFormID TEXT, lockLevel TEXT, primitiveShape TEXT, boundX INTEGER, boundY INTEGER, boundZ INTEGER, rotZ REAL, spawnClass TEXT);
+CREATE TABLE SeventySix_Worldspace(referenceFormID TEXT, x INTEGER, y INTEGER, z INTEGER, locationFormID TEXT, lockLevel TEXT, primitiveShape TEXT, boundX INTEGER, boundY INTEGER, boundZ INTEGER, rotZ INTEGER, spawnClass TEXT);
 CREATE TABLE SeventySix_Cell(cellFormID TEXT, cellEditorID TEXT, cellDisplayName TEXT);
